+++ MISO/Modules/MISO/AppServices/Tests/AppStoreUpdateServiceTests.swift
@@ -6,12 +6,11 @@
 import Foundation
 @testable import MISOModulesAppServices
 import Testing
-import UniformTypeIdentifiers
 
-struct AppStoreUpdateServiceTests {
+struct MISOAppStoreUpdateServiceTests {
 
     @Test
     func `max count limit for release note must be 300`() {
-        #expect(AppStoreUpdateService.RELEASE_NOTE_MAX_COUNT_LIMIT == 300)
+        #expect(MISOAppStoreUpdateService.RELEASE_NOTE_MAX_COUNT_LIMIT == 300)
     }
 }